refactor(stella-protocol): generate the signal-consumer ledger from the tag table
Moves the variant table out of event.rs into event/tags.rs and extends it to
carry a ConsumerPosture per variant, so SIGNAL_CONSUMERS is generated from the
same list as type_tag() and KNOWN_TYPE_TAGS.

The point is where the enforcement lives. The ledger shipped as a
hand-maintained table checked by a test, which was strictly weaker than the
E0004 tripwire the tag list already had. It shipped that way for one reason:
event.rs sat 35 lines under the 1500-line ratchet with no baseline entry, and
the postures are worth roughly a hundred. Splitting the table out removes that
constraint — event.rs drops to 1350 lines and no baseline moved.

Adding an AgentEvent variant without declaring what consumes it is now a
compile error rather than a failing test:

  error[E0004]: non-exhaustive patterns: `&AgentEvent::HaltFired { .. }` not covered
    --> crates/stella-protocol/src/event/tags.rs:95:23

Three LedgerViolation kinds (MissingRow, UnknownTag, DuplicateRow) become
unrepresentable for the real ledger and are deliberately kept: audit_ledger
takes its inputs as parameters so the negative controls can hand it broken
input, and a check that cannot be shown to fail is a claim rather than a
check. The semantic rules it still enforces — issue-citation shape, a named
site for a Behavioral claim, posture/surfaces coherence — are judgements about
a row's content that no macro can hold an author to.

KNOWN_TYPE_TAGS keeps its path via a re-export, so crate::KNOWN_TYPE_TAGS and
every existing citation are unchanged.

Closes #2730
Refs #2701
Refs #2702
